requirements.txt
appnope==0.1.0
click==6.7
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
pygments==2.2.0
pytest-benchmark==3.1.1
pytest-freezegun==0.2.0
simplegeneric==0.8.1
traitlets==4.3.2
